fix player position dislocation

+++ core/src/org/fairysoftw/fairyrogue/actor/PlayerActor.java
@@ -46,22 +46,22 @@ public class PlayerActor extends CreatureActor {
             if (Gdx.input.isKeyJustPressed(Input.Keys.K) ||
                     Gdx.input.isKeyJustPressed(Input.Keys.UP))//up
             {
-                sprite.setY(sprite.getY() + 32);
+                sprite.setY(((int)sprite.getY()) + 32);
             }
             else if (Gdx.input.isKeyJustPressed(Input.Keys.J) ||
                     Gdx.input.isKeyJustPressed(Input.Keys.DOWN))//down
             {
-                sprite.setY(sprite.getY() - 32);
+                sprite.setY(((int)sprite.getY()) - 32);
             }
             else if (Gdx.input.isKeyJustPressed(Input.Keys.H) ||
                     Gdx.input.isKeyJustPressed(Input.Keys.LEFT))//left
             {
-                sprite.setX(sprite.getX() - 32);
+                sprite.setX(((int)sprite.getX()) - 32);
             }
             else if (Gdx.input.isKeyJustPressed(Input.Keys.L) ||
                     Gdx.input.isKeyJustPressed(Input.Keys.RIGHT))//right
             {
-                sprite.setX(sprite.getX() + 32);
+                sprite.setX(((int)sprite.getX()) + 32);
             }
         }
     }
